Bundle kadena-graph and trim closure (#1)
This PR adds a post-processing step to the existing `kadena-graph` output of this flake and bundles the `graph` package down to a single .js file of size 5MB (down from the previous ~200MB). It also removes the unnecessary `prisma` binaries from its environment and further reduces its closure. With these changes in place, the extra closure required for the `kadena-graph` package on top of a Nix environment that already has `nodejs` goes down to 20MB (used to be ~250MB).

* Add a bundled kadena-graph

* Add cwd-extra-migrations to the bundle

* Copy @prisma/client

* Copy all of prisma

* Replace shebang line instead

* Use a subset of prisma-engines

* Shave off further

* Simplify

* Minor simplification

* Create cwd-extra-migrations  folder

* Make the bundled kadena-graph the default

* Inherit packageName from unbundled

* Pass index.js to esbuild instead of idnex.d.ts